moving units on grid does not teleport units
